% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-callback.R
\name{example_callback}
\alias{example_callback}
\title{A function that provides an example dataframe for testing.}
\usage{
example_callback(key)
}
\arguments{
\item{key}{A string used internally (in a specific shiny server module
implementation) to indicate which data to return.}
}
\description{
A function that provides an example dataframe for testing.
}
\examples{
example_callback("county-incidence")

}
