% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-scp.R
\name{get_incidence_data}
\alias{get_incidence_data}
\title{get_incidence_data: Pull cancer incidence data by county for a given set of inputs}
\usage{
get_incidence_data(state, cancer, race, sex, age, stage, year)
}
\arguments{
\item{state}{The US state of interest (character)}

\item{cancer}{The type of cancer (character)}

\item{race}{The race of the population (character)}

\item{sex}{The sex of the population (character)}

\item{age}{The age group of the population (character)}

\item{stage}{The stage of the cancer (character)}

\item{year}{The desired timespan for the data (character)}
}
\value{
A dataframe of age-adjusted cancer incidence rates by county for the input
combination of parameters
}
\description{
Based on user-specified values of cancer, race, sex, age, stage, and
year, collect cancer incidence data by county for the state of interest
}
\examples{
get_incidence_data(
   "wa",
   "all cancer sites",
   "All Races (includes Hispanic)",
   "both sexes",
   "all ages",
   "all stages",
   "latest 5 year average"
)



}
