% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-scp.R
\name{get_incidence_for_all_inputs}
\alias{get_incidence_for_all_inputs}
\title{get_incidence_for_all_inputs: Get incidence for all rows in input_combinations}
\usage{
get_incidence_for_all_inputs(state, input_combinations)
}
\arguments{
\item{state}{The state of interest (character)}

\item{input_combinations}{All possible combinations of input values (dataframe)}
}
\value{
All cancer incidence rates for all counties and all input value combinations
}
\description{
Iterate over all of the rows in input_combinations to get cancer incidence by
county for all possible combinations of input values
}
\examples{
get_incidence_for_all_inputs(
  state,
  all_inputs
)

}
