% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-scp.R
\name{get_input_combinations}
\alias{get_input_combinations}
\title{get_input_combinations: Get all combinations of input parameters}
\usage{
get_input_combinations(
  cancer_types,
  race_options,
  sex_options,
  age_options,
  stage_options,
  year_options
)
}
\arguments{
\item{cancer}{The type of cancer (character)}

\item{race}{The race of the population (character)}

\item{sex}{The sex of the population (character)}

\item{age}{The age group of the population (character)}

\item{stage}{The stage of the cancer (character)}

\item{year}{The desired timespan for the data (character)}
}
\value{
A dataframe of all possible combinations of values of the input variables
}
\description{
Based on the individual options for all of the input variables, generate
all combinations of cancer type, race, sex, age, and year and save them into
a dataframe
}
\examples{
get_input_combinations(
   cancer_types,
   race_options,
   sex_options,
   age_options,
   stage_options
   year_options
)

}
