% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-scp.R
\name{ingest_scp_incidence}
\alias{ingest_scp_incidence}
\title{ingest_scp_incidence: Ingest SCP data by county for all possible inputs}
\usage{
ingest_scp_incidence(
  state,
  cancer_types,
  race_options,
  sex_options,
  age_options,
  stage_options,
  year_options
)
}
\arguments{
\item{state}{The state of interest (character)}

\item{cancer}{The type of cancer (character)}

\item{race}{The race of the population (character)}

\item{sex}{The sex of the population (character)}

\item{age}{The age group of the population (character)}

\item{stage}{The stage of the cancer (character)}

\item{year}{The desired timespan for the data (character)}
}
\value{
The output cancer incidence dataframe to be saved to duckdb for this
combination of parameters
}
\description{
Collect state cancer profiles (SCP) county-level incidence data for all
combinations of cancer types, race, sex, age, stage, and year. Save this data
into the appropriate database
}
\examples{
ingest_scp_incidence(
   "wa",
   cancer_types,
   race_options,
   sex_options,
   age_options,
   stage_options
   year_options
)

}
