% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncarto-file.R
\name{oncarto_file}
\alias{oncarto_file}
\title{Get the paths to files installed with oncarto}
\usage{
oncarto_file(path = NULL)
}
\arguments{
\item{path}{The name of the file. If no argument is provided then
all of the example files will be listed.}
}
\value{
A vector of file paths
}
\description{
This function allows you to quickly access files that are installed with
oncarto.
}
\examples{
oncarto_file()

}
