% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncarto-header.R
\name{oncarto_header}
\alias{oncarto_header}
\title{This function fills in the header of the shiny dashboard}
\usage{
oncarto_header(title, logo_src, logo_href, logo_width, logo_height)
}
\description{
This function fills in the header of the shiny dashboard
}
