% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncarto-sidebar.R
\name{oncarto_sidebar}
\alias{oncarto_sidebar}
\title{This function fills in the sidebar of the shiny dashboard}
\usage{
oncarto_sidebar()
}
\description{
This function fills in the sidebar of the shiny dashboard
}
