% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-scp.R
\name{process_row}
\alias{process_row}
\title{process_row: Get cancer incidence for a given row in the input_combinations df}
\usage{
process_row(state, row)
}
\arguments{
\item{state}{The state of interest}

\item{row}{The current row of the input_combinations df}
}
\value{
The output county-level cancer incidence dataframe for this combination of parameters
}
\description{
Based on a row including a single combination of inputs in the input_combinations df,
get corresponding cancer incidence for the state by county. If the {cancerprof}
API returns a warning, return NULL and print out the warning.
}
\examples{
process_row(
   "wa",
   current_row
)
}
