% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-county-incidence.R
\name{ui_county_incidence}
\alias{ui_county_incidence}
\title{This function specifies the UI for the county incidence tab of the app.}
\usage{
ui_county_incidence(id)
}
\description{
This function specifies the UI for the county incidence tab of the app.
}
