% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-scp.R
\name{write_to_db}
\alias{write_to_db}
\title{write_to_db: Write the dataframe to a database}
\usage{
write_to_db(df, df_name)
}
\arguments{
\item{df}{The type of cancer (character)}

\item{df_name}{The race of the population (character)}
}
\value{
N/A
}
\description{
Save the input dataframe to a database
}
\examples{
write_to_db(
  wa_county_incidence,
  "wa_county_incidence"
)

}
