% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{This function runs the Shiny app packaged within this R package.}
\usage{
run_app(title, logo_src, logo_href, logo_width, logo_height, css, get_data_fn)
}
\arguments{
\item{title}{Title of the application}

\item{logo_src}{PNG or JPEG for the logo}

\item{logo_href}{Website link for the logo}

\item{logo_width}{Width of logo in pixels}

\item{logo_height}{Height of logo in pixels}

\item{css}{Filepath to the .css file that sets aesthetics for the whole app}

\item{get_data_fn}{A function that returns a dataframe according to the name
(a string) specified by the first argument}
}
\description{
This function runs the Shiny app packaged within this R package.
}
